Try harder to find start and end of code blocks
When we see '\n\n```\n' make it a code block start, when we see
'\n```\n\n' make it a code block end. In pm.code_block_start_end(), do
not pm.code_block_start() when we are already inside a code block, and
do not pm.code_block_end() if we are outside a code block.