Simplify code block detection/clean-up code/escape '<'+'>'
Trying to fix and detect the intend of the LLM while streaming is not
feasible. Simplify the detection and deal with the fixing after we have
a complete code block or paragraph and do this in a future commit and
leave it for now.

Clean-up the pattern code and introduce self.pp_next in the process.

Escape < and > when it is part of a tag. The Textual Markdown likes to
swallow it and text between tags when it is not inside code listings.
The LLM is not aware of it.