Replace scrolling logic
When scrolling follows stream we relied on key combinations to enable or
disable it. This was implemented to have a reliable mechanism. It is,
however, not ideal and not intuitive. When implementing an auto scroll
with textual.widget.is_vertical_scroll_end, often the Markdown widget
seemed to trick it into believing it is not at the end, even though it
was. This new implementation seems to work, which leaves an extra line
to consider as the end of the container.