remove `@on(TextArea.Changed)`
When typing text in self.text_area, the input would periodically  not
update for several seconds, making the app unresponsive. This would get
worse when the app is streaming the reply. When we remove it, we get
back a fully reactive application even when streaming. I guess updating
the key bindings on every keystroke is a bad idea. Let's simplify the
logic, leave the key binding active, and just do nothing when the user
hits ^enter when self.text_area.text is empty or the last message is not
a user request.