Smartly handle scroll to end when we change focus
We want to have chat_view scrolled to the end and stay there when we
change focus. Without this change, when changing focus, due to
self.chat_view and self.text_area changing sizes, self.chat_view no
longer is scrolled to the end and stops following the streaming of text.
So, when we were scrolled to the end and change focus, scroll back to
the end.