config: app: add edit settings functionality
We put the Validation class on self.val first. Then we add valuation
methods for the different parts of the edit settings functionality. We
apply the new config structure and make sure we do not create widget ids
containing a '.'. We add actions for removing and adding settings and
hook them up to their respective buttons. We also update the calls to
config_screens by using self instead of cs.