config: settings: add edit settings functionality
We adapt to the new config structure and simplify the store method. We
now only accept an empty value as the default value - which here means,
the setting of the value will not be included in the JSON payload, when
empty. The structure of the settings is now stored in
self.configs[count]["custom"]. This has implications we need to
circumvent should the user decide to manually modify the settings file.
Add methods to remove and add custom settings.