config: validation: simplify, add new validators
We use as much of the functionality of the Input widget to validate the
input. We allow an empty value to have the setting disabled. We add a
new method is_unique_custom to make sure all custom settings are unique.
When we test for an empty value and use is_unique_name or
is_unique_custom, we need to include is_empty, even though we set the
Input field to not allow empty values. Otherwise, for a reason not clear
to me, it will allow empty values, which is not what we want.