settings: endpoint: dismiss new endpoint on cancel
When the user selects to create a new endpoint, default values for a new
endpoint are add to self.config["endpoints"] but are not saved to
endpoints.json. When the user selects "Cancel" after that, the settings
should not appear. Reload the settings from endpoints.json to remove the
settings for the new endpoint.