message: scroll_visible
While we do not want the focused widget to be scrolled into view in most
cases, we do want it to scroll into view after undo/redo and when
editing the content of a message. We also want self.chat_view to scroll
to the end when we append a new request message.