chat: focus
The implementation to choose which message container to focus was
complicated, requiring to feed scroll_visible through many methods and
for some unknown reason would fail when focusing from the side-bar to
the chat-view, focusing the second last message when it should have
focused the last message. Simplify by doing the focusing outside of
message.mount.